Feature/moving dialog modal to UI module
