+++ packages/app/src/crash-dump/components/crash-dump-modal/crash-dump-modal.component.tsx
@@ -7,7 +7,7 @@ import React from "react"
 import { defineMessages } from "react-intl"
 import { intl } from "App/__deprecated__/renderer/utils/intl"
 import { FunctionComponent } from "App/__deprecated__/renderer/types/function-component.interface"
-import ModalDialog from "App/__deprecated__/renderer/components/core/modal-dialog/modal-dialog.component"
+import { ModalDialog } from "App/ui/components/modal-dialog"
 import { ModalSize } from "App/__deprecated__/renderer/components/core/modal/modal.interface"
 import { Size } from "App/__deprecated__/renderer/components/core/button/button.config"
 import Icon from "App/__deprecated__/renderer/components/core/icon/icon.component"
