+++ packages/app/src/overview/components/overview-modals.component.tsx
@@ -35,11 +35,11 @@ import { OverviewModalsTestIds } from "App/overview/components/overview-modals-t
 import Loader from "App/__deprecated__/renderer/components/core/loader/loader.component"
 import { LoaderType } from "App/__deprecated__/renderer/components/core/loader/loader.interface"
 import {
+  ModalDialog,
   RoundIconWrapper,
   ModalMainText,
-} from "App/__deprecated__/renderer/components/core/modal-dialog/modal-dialog-shared"
+} from "App/ui/components/modal-dialog"
 import { Size } from "App/__deprecated__/renderer/components/core/button/button.config"
-import ModalDialog from "App/__deprecated__/renderer/components/core/modal-dialog/modal-dialog.component"
 import { DeviceType } from "@mudita/pure"
 import { IconType } from "App/__deprecated__/renderer/components/core/icon/icon-type"
